Fix for kernel 4.2: don't use 64 bits syscalls/types
The currently Linux kernel version supported by qemu is based on
upstream kernel 4.2 [1].

However, the 64 bits syscalls from musl are not implemented by Hexagon
kernel 4.2. And the time_t type is now 64 bits for all archs on musl
(not only the 64 bits archs), so we also need to force our own type size
here for kernel 4.2.

Furthermore, we need to refrain from using faccessat2, as this was only
added at kernel 5.8.

[1]: https://git.kernel.org/pub/scm/linux/kernel/git/bcain/linux.git/log/?h=bcain/boot_qemu

Signed-off-by: Matheus Tavares Bernardino <matheus.bernardino@oss.qualcomm.com>